Moved the state information link into an accordion that will eventually hold more links.
